Fix an ordering issue around using `migrate`
The `.migrate` method does not actually `tap` anything until you in
turn `tap` it. The sequence of events was:

1. API::Discord.connect would made a new Connection object.
2. That (via a `TWEAK` and `connect`) did an emit on `$!message-source`
3. We then tapped `$!conn.messages`, which in turn tapped the result of
   `migrate`.

By the time 3 happened, however, we'd already emitted to the `supply`
that emits messages from the initial connection. Since the `Supplier`
did not yet have subscribers, the message went to nobody. Use a
`Supplier::Preserving` to make sure that message sent in step 2 is kept
around for step 3.