Fix potential WebSocket deadlock
The use of `.result` in a `whenever` block is blocking. If we get a
message split over many WebSocket frames, then because of the way Supply
backpressure works, we could prevent the handling of those taking place
due to blocking on the `.result`. Even if `Cro::WebSocket` inserts a
defensive `Channel` there at some point in the future, this still makes
far better use of the thread pool by not blocking a thread up.